FastAPI/requirements.txt
aiosmtplib==2.0.2; python_version >= '3.7' and python_version < '4.0'
aiosqlite==0.20.0; python_version >= '3.8'
alembic==1.13.1; python_version >= '3.8'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
asyncpg==0.29.0; python_full_version >= '3.8.0'
bcrypt==4.1.3
blinker==1.8.2; python_version >= '3.8'
certifi==2024.6.2; python_version >= '3.6'
cffi==1.16.0; platform_python_implementation != 'PyPy'
click==8.1.7; python_version >= '3.7'
cloudinary==1.40.0
cryptography==42.0.8
dnspython==2.6.1; python_version >= '3.8'
ecdsa==0.19.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
email-validator==2.1.1; python_version >= '3.8'
fastapi==0.111.0; python_version >= '3.8'
fastapi-cli==0.0.4; python_version >= '3.8'
fastapi-jwt-auth==0.5.0; python_version >= '3.6'
fastapi-limiter==0.1.6; python_version >= '3.9' and python_version < '4.0'
fastapi-mail==1.4.1; python_version < '4.0' and python_full_version >= '3.8.1'
greenlet==3.0.3; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httptools==0.6.1
httpx==0.27.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
jinja2==3.1.4; python_version >= '3.7'
libgravatar==1.0.4
mako==1.3.5; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mdurl==0.1.2; python_version >= '3.7'
orjson==3.10.5; python_version >= '3.8'
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9; python_version >= '3.7'
pyasn1==0.6.0; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pydantic==2.7.4; python_version >= '3.8'
pydantic-core==2.18.4; python_version >= '3.8'
pydantic-settings==2.3.3; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pyjwt==1.7.1
python-dotenv==1.0.1; python_version >= '3.8'
python-jose[cryptography]==3.3.0
python-multipart==0.0.9; python_version >= '3.8'
pyyaml==6.0.1
redis==5.0.6; python_version >= '3.7'
rich==13.7.1; python_full_version >= '3.7.0'
rsa==4.9; python_version >= '3.6' and python_version < '4'
setuptools==70.0.0; python_version >= '3.8'
shellingham==1.5.4; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.30; python_version >= '3.7'
starlette==0.37.2; python_version >= '3.8'
typer==0.12.3; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
ujson==5.10.0; python_version >= '3.8'
urllib3==2.2.2; python_version >= '3.8'
uvicorn[standard]==0.30.1; python_version >= '3.8'
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
